+++ packages/fuselage-ui-kit/src/StaticSelect.js
@@ -6,7 +6,7 @@ import {
 
 import { useBlockContext } from './hooks';
 
-export const StaticSelect = ({
+export const StaticSelect = React.memo(({
   options,
   onChange,
   parser,
@@ -15,17 +15,18 @@ export const StaticSelect = ({
   ...element
 }) => {
   const [{ loading, value, error }, action] = useBlockContext(element, context);
+
   return <SelectFiltered
     error={error}
     value={value}
-    mod-loading={loading}
+    disabled={loading}
     options={options.map((option) => [option.value, parser.text(option.text)])}
     onChange={(value) => action({ target: { value } })}
     placeholder={parser.text(placeholder)} />;
-};
+});
 
 
-export const MultiStaticSelect = ({
+export const MultiStaticSelect = React.memo(({
   context,
   options,
   parser,
@@ -35,9 +36,9 @@ export const MultiStaticSelect = ({
   const [{ loading, value, error }, action] = useBlockContext(element, context);
   return <MultiSelectFiltered
     value={value}
-    mod-loading={loading}
+    disabled={loading}
     error={error}
     options={options.map((option) => [option.value, parser.text(option.text)])}
     onChange={(value) => action({ target: { value } })}
     placeholder={parser.text(placeholder)} />;
-};
+});