+++ packages/fuselage/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.13.0](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.12.0...v0.13.0) (2020-07-14)
+
+
+### Features
+
+* usePosition ([#259](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/259)) ([b395d8c](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/b395d8c2e619d2ba26c5cfbee77275d0fbf9526a))
+
+
+
+
+
+# [0.12.0](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.11.0...v0.12.0) (2020-07-14)
+
+
+### Bug Fixes
+
+* Options Styles ([#260](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/260)) ([410387e](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/410387eb6a049e68f6f6f7391a7e0919c0bdf2ec))
+
+
+### Features
+
+* Margins not only for Box ([#256](https://github.com/RocketChat/Rocket.Chat.Fuselage/issues/256)) ([94e71a8](https://github.com/RocketChat/Rocket.Chat.Fuselage/commit/94e71a80eb37d5145c1d1b2f532d23a782fee07b))
+
+
+
+
+
 # [0.11.0](https://github.com/RocketChat/Rocket.Chat.Fuselage/compare/v0.10.0...v0.11.0) (2020-07-11)